+++ mobile/lib/constants.dart
@@ -6,7 +6,7 @@ const String endpoint = 'http://10.0.2.2:3000';
 const String loremIpsum =
     'Lorem ipsum dolor sit amet, consectetur adipiscing elit sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.';
 
-const Color mainColor = Color(0xFF00242C);
+const Color mainColor = Color(0xFF003E4B);
 const Color secondaryColor = Color(0xFFCDF6FF);
 const Color accentColor = Color(0xFFED5076);
 const Color textButtonColor = Color(0xFF8F00FF);
@@ -23,10 +23,6 @@ const String broadcastChannelTypeFriends = 'FRIENDS';
 const String broadcastTypeFriendRequest = 'FRIEND_REQUEST';
 const String broadcastTypeFriendAccepted = 'FRIEND_REQUEST_ACCEPTED';
 const String broadcastTypeFriendDeleted = 'FRIEND_DELETED';
-
-const String broadcastTypeGoalUpdate = 'GOAL_UPDATED';
-const String broadcastTypeGoalSetMilestones = 'GOAL_SET_MILESTONES';
-const String broadcastTypeGoalDelete = 'GOAL_DELETED';
 
 const String broadcastChannelTypeAlerts = 'ALERTS';
 const String broadcastTypeNewAlert = 'NEW_ALERT';