+++ mobile/lib/models/index.dart
@@ -6,3 +6,6 @@ export './friendship.dart';
 export './notification.dart';
 export './image.dart';
 export './event.dart';
+export './reaction.dart';
+export './comment.dart';
+export './view.dart';