+++ mobile/lib/screens/goal/goal_milestones_view.dart
@@ -18,10 +18,14 @@ class _GoalMilestonesViewState
   @override
   Future afterFirstBuild(BuildContext context) async {
     setIsLoading(true);
-    await context
-        .read<SingleGoalProvider>()
-        .populate(ModalRoute.of(context)!.settings.arguments as String);
+    await context.read<SingleGoalProvider>().populate(store.state.user!,
+        ModalRoute.of(context)!.settings.arguments as String);
     setIsLoading(false);
+
+    subscribeToChannel('GOAL', viewModel.model!.goal.id, (message) {
+      viewModel.processMessage(message);
+      setState(() {});
+    });
   }
 
   @override
@@ -172,10 +176,8 @@ class _GoalMilestonesViewState
       return;
     }
 
-    final milestones = viewModel.model!.goal.milestones.toList();
-    milestones.remove(milestone);
     setIsLoading(true);
-    viewModel.setMilestones(milestones).then((value) {
+    viewModel.deleteMilestone(milestone).then((value) {
       setIsLoading(false);
       setState(() {});
     });